+++ src/app/typings/index.ts
@@ -1,1 +1,2 @@
 export * from './user/user.interfaces';
+export * from './auth/auth.interfaces';